SFW update helper transform to a class.
